services/cart/requirements.txt
flask
flask-cors
flask-swagger-ui
requests
python-dotenv
jsonschema
